Restructure nvidia-smi fixture usage and example test
This uses return values from the fixtures, rather than attaching things to the
Pytest module, which is a bit more idiomatic.  It also prefers shorter names.

Then we change around a single test to serve as an example.
This does a single call and avoids the indirection of a function definition.